[Refactor] Move classes into ColumnCore and ChunkCore
