+++ be/src/bench/shuffle_chunk_bench.cpp
@@ -20,8 +20,8 @@
 
 #include "column/chunk.h"
 #include "column/column_helper.h"
+#include "column/segmented_chunk.h"
 #include "column/vectorized_fwd.h"
-#include "storage/chunk_helper.h"
 #include "types/logical_type.h"
 #include "types/type_descriptor.h"
