+++ be/src/column/chunk_extra_data.h
@@ -15,7 +15,6 @@
 #pragma once
 
 #include "column/chunk.h"
-#include "serde/column_array_serde.h"
 #include "types/type_descriptor.h"
 
 namespace starrocks {
@@ -56,11 +55,6 @@ public:
 
     size_t memory_usage() const;
     size_t bytes_usage(size_t from, size_t size) const;
-
-    int64_t max_serialized_size(const int encode_level = 0);
-    StatusOr<uint8_t*> serialize(uint8_t* buff, bool sorted = false, const int encode_level = 0);
-    StatusOr<const uint8_t*> deserialize(const uint8_t* buff, const uint8_t* end, bool sorted = false,
-                                         const int encode_level = 0);
 
     static ChunkExtraColumnsData* as_raw(const ChunkExtraDataPtr& extra_data);
