+++ be/src/storage/chunk_helper.h
@@ -18,6 +18,7 @@
 #include <memory>
 
 #include "column/chunk.h"
+#include "column/segmented_chunk.h"
 #include "column/vectorized_fwd.h"
 #include "storage/olap_common.h"
 #include "tablet_schema.h"
@@ -163,70 +164,6 @@ private:
     // so incremental calculation is used to avoid becoming a performance bottleneck.
     size_t _mem_usage = 0;
     bool _finalized = false;
-};
-
-class SegmentedColumn final : public std::enable_shared_from_this<SegmentedColumn> {
-public:
-    SegmentedColumn(const SegmentedChunkPtr& chunk, size_t column_index);
-    SegmentedColumn(Columns columns, size_t segment_size);
-    ~SegmentedColumn() = default;
-
-    ColumnPtr clone_selective(const uint32_t* indexes, uint32_t from, uint32_t size);
-    ColumnPtr materialize() const;
-
-    bool is_nullable() const;
-    bool has_null() const;
-    size_t size() const;
-    void upgrade_to_nullable();
-    size_t segment_size() const;
-    size_t num_segments() const;
-    Columns columns() const;
-
-private:
-    SegmentedChunkWeakPtr _chunk; // The chunk it belongs to
-    size_t _column_index;         // The index in original chunk
-    const size_t _segment_size;
-
-    Columns _cached_columns; // Only used for SelectiveCopy
-};
-
-// A big-chunk would be segmented into multi small ones, to avoid allocating large-continuous memory
-// It's not a transparent replacement for Chunk, but must be aware of and set a reasonale chunk_size
-class SegmentedChunk final : public std::enable_shared_from_this<SegmentedChunk> {
-public:
-    SegmentedChunk(size_t segment_size);
-    ~SegmentedChunk() = default;
-
-    static SegmentedChunkPtr create(size_t segment_size);
-
-    void append_column(ColumnPtr column, SlotId slot_id);
-    void append_chunk(const ChunkPtr& chunk, const std::vector<SlotId>& slots);
-    void append_chunk(const ChunkPtr& chunk);
-    void append(const SegmentedChunkPtr& chunk, size_t offset);
-    void build_columns();
-
-    SegmentedColumnPtr get_column_by_slot_id(SlotId slot_id);
-    const SegmentedColumns& columns() const;
-    SegmentedColumns& columns();
-    size_t num_segments() const;
-    const std::vector<ChunkPtr>& segments() const;
-    std::vector<ChunkPtr>& segments();
-    ChunkUniquePtr clone_empty(size_t reserve);
-
-    size_t segment_size() const;
-    void reset();
-    size_t memory_usage() const;
-    size_t num_rows() const;
-    Status upgrade_if_overflow();
-    Status downgrade();
-    bool has_large_column() const;
-    void check_or_die();
-
-private:
-    std::vector<ChunkPtr> _segments;
-    SegmentedColumns _columns;
-
-    const size_t _segment_size;
 };
 
 class ExprContext;
