+++ be/test/column/CMakeLists.txt
@@ -32,7 +32,9 @@ set(COLUMN_TEST_FILES
     const_nullable_core_test.cpp
     nullable_column_core_test.cpp
     decimalv3_json_core_test.cpp
+    append_with_mask_core_test.cpp
     datum_convert_core_test.cpp
+    column_helper_time_core_test.cpp
     json_converter_core_test.cpp
     adaptive_nullable_column_core_test.cpp
     column_view_base_core_test.cpp
@@ -71,6 +73,7 @@ target_link_libraries(column_test ${COLUMN_TEST_LINK_LIBS} gtest_main)
 set(CHUNK_TEST_FILES
     ../base/cxa_throw_wrap.cpp
     chunk_core_test.cpp
+    segmented_chunk_core_test.cpp
     field_test.cpp
     schema_test.cpp
 )
